+++ Restaurant.API/Program.cs
@@ -1,59 +1,37 @@
 using Restaurant.Infrastructure;
 using Restaurant.Application;
-using Microsoft.OpenApi.Models;
 using Restaurant.API.Middleware;
 using FluentValidation;
+using Restaurant.Application.DTOs.Auth;
+using Restaurant.API.Utilities;
 
 var builder = WebApplication.CreateBuilder(args);
 
 // Add services to the container.
+builder.Configuration.AddEnvironmentVariables();
+builder.RegisterAuthorization();
 builder.Services.AddControllers();
+
+builder.Services.Configure<JwtSettings>(options => {
+    options.Key = Environment.GetEnvironmentVariable("JWT_KEY") ?? throw new ArgumentNullException("JWT_KEY", "JWT_KEY environment variable is not set");
+    options.Issuer = Environment.GetEnvironmentVariable("JWT_ISSUER") ?? throw new ArgumentNullException("JWT_ISSUER", "JWT_ISSUER environment variable is not set");
+    options.Audience = Environment.GetEnvironmentVariable("JWT_AUDIENCE") ?? throw new ArgumentNullException("JWT_AUDIENCE","JWT_ISSUER environment variable is not set"); ;
+
+    if (int.TryParse(Environment.GetEnvironmentVariable("JWT_ACCESS_TOKEN_EXPIRY"), out int accessExpiry))
+        options.AccessTokenExpiryMinutes = accessExpiry;
+
+    if (int.TryParse(Environment.GetEnvironmentVariable("JWT_REFRESH_TOKEN_EXPIRY_DAYS"), out int refreshExpiry))
+        options.RefreshTokenExpiryInDays = refreshExpiry;
+});
+
 builder.Services.AddValidatorsFromAssembly(
     typeof(Restaurant.Application.AssemblyMarker).Assembly);
 builder.Services.AddAutoMapper(AppDomain.CurrentDomain.GetAssemblies());
 builder.Services.AddApplicationServices();
 builder.Services.AddInfrastructureServices(builder.Configuration);
-builder.Services.AddAuthentication();
-builder.Services.AddSwaggerGen(c =>
-{
-    c.AddSecurityDefinition("Bearer", new OpenApiSecurityScheme
-    {
-        Description = "JWT Authorization header using the Bearer scheme.",
-        Name = "Authorization",
-        In = ParameterLocation.Header,
-        Type = SecuritySchemeType.ApiKey,
-        Scheme = "Bearer",
-    });
-
-    c.AddSecurityRequirement(new OpenApiSecurityRequirement
-    {
-        {
-            new OpenApiSecurityScheme
-            {
-                Reference = new OpenApiReference
-                {
-                    Type = ReferenceType.SecurityScheme,
-                    Id = "Bearer",
-                },
-            },
-            Array.Empty<string>()
-        },
-    });
-
-    c.SwaggerDoc("v1", new OpenApiInfo
-    {
-        Version = "v1",
-        Title = "Restaurant.WebApi",
-        Description = "A Web API for managing restaurant operations",
-    });
-
-    var xmlFile = $"{System.Reflection.Assembly.GetExecutingAssembly().GetName().Name}.xml";
-    var xmlPath = Path.Combine(AppContext.BaseDirectory, xmlFile);
-    c.IncludeXmlComments(xmlPath);
-});
 builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
-
+builder.AddSwaggerDoc();
 var app = builder.Build();
 
 app.UseMiddleware<ExceptionMiddleware>();