+++ Restaurant.Application/DTOs/Auth/RefreshTokenRequest.cs
@@ -1,0 +1,7 @@
+﻿namespace Restaurant.Application.DTOs.Auth
+{
+    public class RefreshTokenRequest
+    {
+        public required string RefreshToken { get; set; }
+    }
+}