+++ Restaurant.Application/DTOs/Dishes/DishDetailsDto.cs
@@ -7,14 +7,10 @@ public class DishDetailsDto : DishDto
     public string? Carbohydrates { get; set; }
     
     public string? Description { get; set; }
-
-    public string? DishType { get; set; }
     
     public string? Fats { get; set; }
 
     public string? Proteins { get; set; }
-
-    public string? State { get; set; }
 
     public string? Vitamins { get; set; }
 }