+++ Restaurant.Application/Interfaces/IAuthService.cs
@@ -6,5 +6,9 @@ namespace Restaurant.Application.Interfaces
     public interface IAuthService
     {
         Task<string> RegisterUserAsync(RegisterDto request);
+
+        Task<TokenResponseDto> LoginAsync(LoginDto request);
+
+        Task<TokenResponseDto> RefreshTokenAsync(string refreshToken);
     }
 }