+++ Restaurant.API/CHANGELOG.md
@@ -4,7 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.7.1] - 2025-04-14
+## [0.9.1] - 2025-04-15
+### Added
+
+- Migrated `GET /dishes` endpoint logic
+- Added support for new environment variables
+- Implemented login and refresh token functionality
+- Integrated Swagger XML documentation into controllers
+
+## [0.8.1] - 2025-04-14
 ### Added
 
 - Migrated `GET /dishes/{id}` endpoint to microservice architecture.